Update reference to libethard to libudpard
As a result of decisions made regarding the utility of using CAN over
ethernet, it is decided the protocols should be specific to the protocol
used. i.e. TCP/IP vs. UDP. Because this package is specific to the UDP
protocol, all references to libethard should be replaced with libudpard
as it may be ambiguous if this could be used for tcp/ip.