Restart RX Session if out-of-order multiframe transfer received (#20)
During on-target testing, it was found that if a frame is received out
of order during a multiframe transfer, then libudpard cannot receive any
subsequent transfers for any port IDs because it reports an out of
memory error. This is because the memory allocated for the out-of-order
multiframe transfer is never freed. Since out-of-order multiframe
transfers are not currently supported, the best course of action is drop
the entire payload and require the user to resend it.