New API and static analysis (#31)
1. Update the API according to the new design principle that the library
is to be used with a third-party UDP/IP+IGMP stack.

> Q: Could you be a bit more specific about the multiplexing that you
saw in the old version? Do you mean redundant interfaces?
> A: No, I mean the socket layer (or the lack-of-socket layer, as Erik
puts it) already demultiplexes incoming datagrams per subject. The
original design joins the demultiplexed datagram feeds into one feed
which is then demultiplexed back through AVL tree lookup. This is
suboptimal. The new design eliminates one (of the two) AVL lookups from
the RX pipeline. To subscribe to a subject, you would call
`udpardRxSubscribtionInit` and then create a socket for this
subscription. The socket and the subscription will exist side-by-side.

2. Set up static analysis and verification:
 - Clang-Tidy and SonarCloud with a large subset of MISRA rules enabled.
 - Build for i386, AMD64, ARM, AVR.
 - Run tests on i386 and AMD64 using both GCC and Clang with coverage.

3. Switch from Catch2 to Google Test.