Implement subscription API functions (#46)
It is now possible to subscribe to subjects and receive data from them.
The only missing bit is the RPC services API, all other functionality of
the library is already implemented and can be used.

This changeset also brings `udpardGather` that, well, gathers a
fragmented payload buffer into one contiguous chunk of memory that can
be passed into Nunavut. Later we should amend Nunavut to support
fragmented buffers directly to eliminate extra data copying and memory
utilization.

This changeset also fixes an internal documentation error that provided
an incorrect estimation of the worst-case recursion depth.