Add RPC service support to the RX pipeline (#47)
This is the final feature.

This changeset slightly modifies the error handling behavior of
`udpardRxSubscriptionReceive` such that the datagram is deallocated if
the passed arguments are invalid, unless `self` is also invalid. This is
needed to prevent memory leak when the function is invoked incorrectly.
An alternative solution is to state it explicitly in the API that the
caller is required to free the datagram memory if
`-UDPARD_ERROR_ARGUMENT` is returned, but this convention is
comparatively fragile. A new test is added for this.

`UdpardRxRPC` is renamed into `UdpardRxRPCPort` for clarity.

`UdpardRxRPCDispatcher` is extended with `local_node_id` to facilitate
destination address check when accepting frames.